requirements.txt
requests==2.31.0
google-genai
python-dotenv==1.0.1
rich==13.7.1
